# openapi.yaml
openapi: 3.0.1
info:
  title: Airport Information Service
  description: A service that returns airport names, codes, lat/long, and urls.
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /nearestAirports/{lat}/{long}/{count}:
    get:
      operationId: findNearestAirports
      summary: Find the nearest airports by latitude and longitude
      parameters:
        - in: path
          name: lat
          schema:
            type: number
            format: double
          required: true
          description: The latitude of the location
        - in: path
          name: long
          schema:
            type: number
            format: double
          required: true
          description: The longitude of the location
        - in: path
          name: count
          schema:
            type: integer
          required: true
          description: The number of nearest airports to return
      responses:
        '200':
          description: A list of the nearest airports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Airport'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /searchByName/{cityName}:
    get:
      summary: Find airports with similar names based on a town/city
      operationId: searchAirportsByName
      parameters:
        - in: path
          name: cityName
          schema:
            type: string
          required: true
          description: The name of the town/city
      responses:
        '200':
          description: A list of airports with similar names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Airport'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /searchByIdent/{ident}:
    get:
      summary: Find airports with similar identifiers based on a provided code
      operationId: searchAirportsByIdent
      parameters:
        - in: path
          name: ident
          schema:
            type: string
          required: true
          description: The airport identifier or code
      responses:
        '200':
          description: A list of airports with similar identifiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Airport'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /flightPlan:
    post:
      summary: Generate a flight plan link
      operationId: generateFlightPlanLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                coordinates:
                  type: array
                  items:
                    type: object
                    properties:
                      lat:
                        type: number
                      long:
                        type: number
                    required:
                      - lat
                      - long
                  example:
                    - lat: -41.023798
                      long: 147.416791
                    - lat: -41.54529953
                      long: 147.214004517
                    - lat: -41.336700439453125
                      long: 148.28199768066406
              required:
                - coordinates
      responses:
        200:
          description: Flight plan link generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  flightPlanLink:
                    type: string
                    format: uri
                    example: https://skyvector.com/?fpl=410128S1472500E%20413251S1471248E%20412012S1481655E
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid input provided.

components:
  schemas:
    Airport:
      type: object
      properties:
        name:
          type: string
          description: The airport name
        ident:
          type: string
          description: The airport ident code
        lat:
          type: number
          format: double
          description: The airport latitude
        long:
          type: number
          format: double
          description: The airport longitude
        url:
          type: string
          format: uri
          description: The airport URL
    Error:
      type: object
      properties:
        message:
          type: string
          description: A description of the error